uhd: Remove explicit check for devices with EDGE support.
Now we already have 4 devices (B200, B210, UmTRX, LimeSDR) and it's getting
crowded. We now have a generic check which complains if a configuration is not
found in the offset table. This should be enough for all practical reasons.
If we really want to complain about EDGE support, then we can just check
'edge' flag and change the log message accordingly. I don't see a real need
for this, though.